pyproject.toml
[project]
name = "buda"
version = "0.1.0"
description = "Buda.com API client for REST and WebSocket APIs with sync and async support."
readme = { file = "README.md", content-type = "text/markdown" }
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.13,<3.14.0"

authors = [
    { name = "Matías Schiaffino Tyrer", email = "matias.scht@gmail.com" }
]

dependencies = [
    "httpx>=0.28.1,<1.0.0",
    "pydantic-settings>=2.13.1,<3.0.0",
    "tenacity>=9.1.4,<10.0.0",
    "websockets>=16.0,<17.0.0",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Typing :: Typed",
    "Framework :: Pydantic :: 2",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "ipykernel>=7.2.0,<8.0.0",
    "ipywidgets>=8.1.8,<9.0.0",
    "notebook>=7.5.5,<8.0.0",
    "pyright>=1.1.408,<2.0.0",
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.15.8,<1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["buda"]
